fix sql

+++ data/sql/postgres/common/tables/dimensions.sql
@@ -1,5 +1,5 @@
 create table dimensions (
-    id tinyint not null,
+    id smallint not null,
     name varchar(100) not null,
 
     constraint pk_dimensions
